C++11: enum -> enum class

+++ Include/xLib/Sync/Platform/Unix/Thread_unix.inl
@@ -146,7 +146,7 @@ Thread::_setPriority_impl(
     }
 
     sched_param param = {0};
-    param.sched_priority = a_priority;
+    param.sched_priority = static_cast<int>(a_priority);
 
     int_t iRv = ::pthread_setschedparam(id(), SCHED_FIFO, &param);
     xTEST_EQ_MSG(0, iRv, NativeError::format( static_cast<ulong_t>(iRv) ));
